pyproject.toml
[tool.poetry]
name = "hal9000"
version = "0.1.0"
description = "Hal9000 the Discord bot"
authors = ["Aaron West <aphexlog@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
venv = ".venv"
venvPath = "."
